#31 refactor: AppBar_UI.dart 파일에 빠져있던 MyAppBar class 추가 및 기존의 AppBar.dart 파일 제거
