+++ lib/AfterLogin.dart
@@ -1,4 +1,4 @@
-import 'package:academy_manager/AppBar.dart';
+import 'package:academy_manager/UI/AppBar_UI.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'NoticeDetail.dart';  // NoticeDetail 화면을 import
