+++ lib/AppSettings.dart
@@ -1,4 +1,4 @@
-import 'package:academy_manager/AppBar.dart';  // AppBar.dart 파일에서 MyAppBar를 import
+import 'package:academy_manager/UI/AppBar_UI.dart';  // AppBar.dart 파일에서 MyAppBar를 import
 import 'package:academy_manager/MyDio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
