+++ lib/ScoreGraph.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:academy_manager/AppBar.dart';
+import 'package:academy_manager/UI/AppBar_UI.dart';
 import 'package:fl_chart/fl_chart.dart';
 
 class ScoreGraph extends StatefulWidget {
